Checking in the backend changes for the choice of crossfade fade method... Either linear or parabolic.
Added TSoundStretcher which is able to be used as a sample rate converter.  At the moment, it's a trivial implementation, but I will enhance it later to interpole between samples.
Added a sampleRate parameter to CSound::mixSound so it can call on TSoundStretcher to convert the two sources to be mixed to the sample sample rate.